+++ gatsby-browser.js
@@ -2,4 +2,4 @@ import python from 'codemirror/mode/python/python' // eslint-disable-line no-unu
 
 // This doesn't have to be here – but if we do import Juniper here, it's already
 // preloaded and cached when we dynamically import it in code.js.
-import Juniper from './src/components/JuniperComponent' // eslint-disable-line no-unused-vars
+import Juniper from './src/components/Juniper' // eslint-disable-line no-unused-vars
