+++ src/markdown.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import rehypeReact from 'rehype-react';
 
-import ExerciseComponent from './components/ExerciseComponent';
-import { TextBlockComponent } from './components/TextBlockComponent';
+import Exercise from './components/Exercise';
+import { TextBlock } from './components/TextBlock';
 import CodeBlock from './components/CodeBlock';
-import { LinkComponent } from './components/LinkComponent';
+import { Link } from './components/Link';
 
 export const renderAst = new rehypeReact({
     createElement: React.createElement,
     components: {
-        exercise: ExerciseComponent,
-        textblock: TextBlockComponent,
+        exercise: Exercise,
+        textblock: TextBlock,
         codeblock: CodeBlock,
-        a: LinkComponent,
+        a: Link,
     },
 }).Compiler;
