+++ src/pages/404.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '../templates/Layout';
+import Layout from '../components/Layout';
 
 export default () => {
     return (
